# output/architecture/openapi.yaml
openapi: 3.1.2
...
info:
  title: Checkout sandbox API
  version: 1.0.0
  summary: Contrato HTTP del checkout invitado asíncrono
  description: |
    Contrato contract-first de Etapa 3 consumido por la construcción de Etapa 5.
    El servidor es autoridad de precio, moneda, stock, cotización y estado.
    Los datos completos de tarjeta nunca forman parte de esta API. Los valores
    efímeros de pago y aceptación sólo aparecen como campos de escritura sin
    ejemplos, defaults ni persistencia.
  x-status: STAGE_5_LOCAL_FAKE_VERIFIED
  x-source-baseline:
    requirements: 33 RF identities and 28 RNF identities
    acceptanceCriteria: 45
    scenarios: 51
    errors: 24
    dataFields: 72
    uat: 48
  x-design-decisions:
    oasVersion: OpenAPI 3.1.2 for reversible toolchain interoperability
    customerRoute: One nested PUT route under the checkout; no duplicate top-level write route
    stockRoute: Kept as a separate nominal read resource for the challenge rubric
    webhookRoute: One singular optional P1 route
    capability: HttpOnly cookie; raw value never enters JSON, URLs, logs or examples
    paymentSubmission: Every locally accepted payment submission and its semantic replay return the same 202 body and Location
    asyncOutcomes: Provider uncertainty and later incidents are represented by the local transaction resource
    localStateVocabulary: E0-E2 values NOT_SENT and ACKNOWLEDGED are preserved
...
servers:
  - url: https://checkout.example.invalid
...
security: []
...
paths:
...
  /api/v1/checkouts:
    post:
      operationId: createCheckout
      tags: [Checkouts]
      summary: Crear checkout, cotización y capability
      description: |
        Crea el recurso local sincrónicamente. La capability cruda se entrega
        exclusivamente mediante Set-Cookie HttpOnly; nunca aparece en JSON.
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCheckoutRequest' }
            examples:
              oneProduct:
                value: { productId: PRODUCT_ID_OPAQUE }
      x-api-id: API-04
      x-status: IMPLEMENTED
      x-browser-security:
        csrf: Same-origin plus Origin and Fetch Metadata validation
        cookie: HttpOnly; Secure; SameSite=Strict; Path=/api/v1
        cache: No shared storage of Set-Cookie responses
      x-trace:
        requirements: [RF-02, RF-13, RF-24, RF-32]
        acceptanceCriteria: [AC-US-02-01, AC-US-04-01]
        scenarios: [SC-US-02-01, SC-US-04-01, SC-EN-11-03]
        uat: [UAT-17, UAT-21]
        data: [DAT-01, DAT-06, DAT-07, DAT-10, DAT-11, DAT-12, DAT-13, DAT-14, DAT-15, DAT-16, DAT-17, DAT-18, DAT-19, DAT-20, DAT-21, DAT-22]
        errors: [ERR-01, ERR-02, ERR-04, ERR-05, ERR-06, ERR-19, ERR-20]
        uxFlows: [UXF-01, UXF-03]
      responses:
        '201':
          description: Checkout creado
          headers:
            Location: { $ref: '#/components/headers/Location' }
            Set-Cookie: { $ref: '#/components/headers/CheckoutCapabilityCookie' }
            ETag: { $ref: '#/components/headers/StrongETag' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            Vary: { $ref: '#/components/headers/VaryOrigin' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutCreatedResponse' }
              examples:
                created:
                  value:
                    checkoutId: CHECKOUT_ID_OPAQUE
                    status: DRAFT
                    version: 1
                    quote:
                      quoteId: QUOTE_ID_OPAQUE
                      version: 1
                      productId: PRODUCT_ID_OPAQUE
                      quantity: 1
                      subtotal: { amountInCents: 2500000, currency: COP }
                      baseFee: { amountInCents: 200000, currency: COP }
                      deliveryFee: { amountInCents: 500000, currency: COP }
                      total: { amountInCents: 3200000, currency: COP }
                      expiresAt: '2026-01-01T00:15:00Z'
                    expiresAt: '2026-01-01T00:30:00Z'
        '400': { $ref: '#/components/responses/RequestMalformed' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '404': { $ref: '#/components/responses/ProductNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/FieldInvalid' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/checkouts/{checkoutId}:
    get:
      operationId: getCheckout
      tags: [Checkouts]
      summary: Recuperar el checkout canónico
      description: |
        Lectura local sin efectos ni consulta al proveedor. Inexistencia,
        capability inválida y relación ajena producen el mismo 404.
      security:
        - CheckoutCapability: []
      parameters:
        - $ref: '#/components/parameters/CheckoutId'
        - $ref: '#/components/parameters/OptionalOrigin'
        - $ref: '#/components/parameters/CorrelationId'
      x-api-id: API-05
      x-status: IMPLEMENTED
      x-trace:
        requirements: [RF-13, RF-32]
        acceptanceCriteria: [AC-US-09-01, AC-US-09-02, AC-US-09-05]
        scenarios: [SC-US-09-01, SC-US-09-03, SC-EN-11-02, SC-EN-11-04]
        uat: [UAT-17, UAT-25, UAT-26, UAT-28]
        data: [DAT-11, DAT-12, DAT-13, DAT-15, DAT-16, DAT-17, DAT-18, DAT-19, DAT-20, DAT-21, DAT-22, DAT-29, DAT-32, DAT-42, DAT-43, DAT-44, DAT-45, DAT-46, DAT-47, DAT-48, DAT-49, DAT-50]
        errors: [ERR-02, ERR-03, ERR-08, ERR-20]
        uxFlows: [UXF-12, UXF-13]
      responses:
        '200':
          description: Checkout autorizado
          headers:
            ETag: { $ref: '#/components/headers/StrongETag' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutResponse' }
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '404': { $ref: '#/components/responses/CheckoutHidden' }
        '410': { $ref: '#/components/responses/CheckoutExpired' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/checkouts/{checkoutId}/customer:
    put:
      operationId: replaceCheckoutCustomer
      tags: [Customers]
      summary: Guardar o reemplazar el cliente del checkout
      description: Reemplazo idempotente protegido por capability, relación y ETag fuerte.
      security:
        - CheckoutCapability: []
      parameters:
        - $ref: '#/components/parameters/CheckoutId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceCustomerRequest' }
      x-api-id: API-06
      x-status: IMPLEMENTED
      x-decision-status: ASSUMED_FIELDS_FROM_DEC-07
      x-trace:
        requirements: [RF-21, RF-27, RF-32]
        acceptanceCriteria: [AC-US-03-03]
        scenarios: [SC-US-03-01, SC-EN-11-02, SC-EN-11-03]
        uat: [UAT-17, UAT-19, UAT-44]
        data: [DAT-11, DAT-13, DAT-15, DAT-41, DAT-42, DAT-43, DAT-44]
        errors: [ERR-01, ERR-02, ERR-03, ERR-05, ERR-08, ERR-09, ERR-19, ERR-20]
        uxFlows: [UXF-05, UXF-12]
      responses:
        '200':
          description: Cliente guardado
          headers:
            ETag: { $ref: '#/components/headers/StrongETag' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerResponse' }
        '400': { $ref: '#/components/responses/RequestMalformed' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '404': { $ref: '#/components/responses/CheckoutHidden' }
        '410': { $ref: '#/components/responses/CheckoutExpired' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/FieldInvalid' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/checkouts/{checkoutId}/transactions:
    post:
      operationId: submitCheckoutTransaction
      tags: [Payments, Transactions]
      summary: Aceptar una intención de pago
      description: |
        If-Match, Idempotency-Key, capability y relación son obligatorios.
        Reserva, idempotencia y transacción local PENDING quedan durables antes
        del I/O externo. Toda intención local aceptada y todo replay semántico
        devuelven el mismo 202, body y Location originales. El estado vigente se
        consulta mediante GET; no se deduce del status HTTP de este POST.
      security:
        - CheckoutCapability: []
      parameters:
        - $ref: '#/components/parameters/CheckoutId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentSubmissionRequest' }
      x-api-id: API-09
      x-status: IMPLEMENTED
      x-stable-accepted-response:
        status: 202
        replayInvariantFields: [transactionId, statusUrl, submissionState, acceptedAt, Location]
        perRequestField: X-Correlation-Id
      x-trace:
        requirements: [RF-07, RF-08, RF-24, RF-29, RF-30, RF-32]
        acceptanceCriteria: [AC-US-05-01, AC-US-05-02, AC-US-05-03, AC-US-05-04, AC-US-05-05, AC-US-10-01, AC-US-10-02, AC-US-10-03, AC-US-10-04]
        scenarios: [SC-US-05-01, SC-US-05-02, SC-US-05-03, SC-US-05-04, SC-US-10-01, SC-US-10-02, SC-US-10-03]
        uat: [UAT-01, UAT-04, UAT-05, UAT-06, UAT-07, UAT-08, UAT-22, UAT-23, UAT-24, UAT-45, UAT-48]
        data: [DAT-11, DAT-13, DAT-15, DAT-16, DAT-17, DAT-22, DAT-23, DAT-24, DAT-26, DAT-27, DAT-28, DAT-29, DAT-30, DAT-31, DAT-32, DAT-33, DAT-34, DAT-35, DAT-36, DAT-37, DAT-38, DAT-39, DAT-40, DAT-57]
        errors: [ERR-01, ERR-02, ERR-03, ERR-05, ERR-06, ERR-07, ERR-08, ERR-09, ERR-10, ERR-11, ERR-12, ERR-13, ERR-14, ERR-18, ERR-19, ERR-20, ERR-23, ERR-24]
        uxFlows: [UXF-01, UXF-06, UXF-07, UXF-08, UXF-09]
      responses:
        '202':
          description: Intención local aceptada; usar Location para consultar
          headers:
            Location: { $ref: '#/components/headers/Location' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentSubmissionResponse' }
              examples:
                accepted:
                  value:
                    transactionId: TRANSACTION_LOCAL_ID
                    statusUrl: /api/v1/transactions/TRANSACTION_LOCAL_ID
                    submissionState: ACCEPTED
                    acceptedAt: '2026-01-01T00:00:00Z'
        '400': { $ref: '#/components/responses/RequestMalformed' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '404': { $ref: '#/components/responses/CheckoutHidden' }
        '409': { $ref: '#/components/responses/PaymentConflict' }
        '410': { $ref: '#/components/responses/CheckoutExpired' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/FieldInvalid' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/transactions/{transactionId}:
    get:
      operationId: getTransaction
      tags: [Transactions]
      summary: Consultar el estado local canónico
      description: |
        Sólo lee estado local. Nunca llama al proveedor, finaliza, libera,
        reserva ni crea otro pago. PENDING, APPROVED, DECLINED, VOIDED y ERROR
        se representan con 200; UNKNOWN es una fase de despacho, no un final.
      security:
        - CheckoutCapability: []
      parameters:
        - $ref: '#/components/parameters/TransactionId'
        - $ref: '#/components/parameters/OptionalOrigin'
        - $ref: '#/components/parameters/CorrelationId'
      x-api-id: API-10
      x-status: IMPLEMENTED
      x-trace:
        requirements: [RF-09, RF-13, RF-26, RF-31, RF-32]
        acceptanceCriteria: [AC-US-06-01, AC-US-06-02, AC-US-06-03, AC-US-09-03, AC-US-09-04]
        scenarios: [SC-US-06-01, SC-US-06-02, SC-US-09-02]
        uat: [UAT-03, UAT-23, UAT-27, UAT-34, UAT-35, UAT-40, UAT-41, UAT-42]
        data: [DAT-11, DAT-15, DAT-23, DAT-24, DAT-29, DAT-30, DAT-31, DAT-32, DAT-33, DAT-34, DAT-35, DAT-51, DAT-52, DAT-69]
        errors: [ERR-02, ERR-03, ERR-12, ERR-13, ERR-14, ERR-18, ERR-20, ERR-21, ERR-22]
        uxFlows: [UXF-09, UXF-10, UXF-11, UXF-12, UXF-13]
      responses:
        '200':
          description: Recurso de transacción local, terminal o no terminal
          headers:
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
            Retry-After: { $ref: '#/components/headers/OptionalRetryAfter' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
              examples:
                unknown:
                  value:
                    transactionId: TRANSACTION_LOCAL_ID
                    checkoutId: CHECKOUT_ID_OPAQUE
                    checkoutStatus: PAYMENT_PENDING
                    paymentStatus: PENDING
                    dispatchPhase: UNKNOWN
                    providerStatus: null
                    reservationStatus: ACTIVE
                    integrityStatus: OK
                    recoveryCode: PROVIDER_OUTCOME_UNKNOWN
                    statusUrl: /api/v1/transactions/TRANSACTION_LOCAL_ID
                    allowedActions: [QUERY, WAIT, RETURN_TO_PRODUCT]
                    acceptedAt: '2026-01-01T00:00:00Z'
                    updatedAt: '2026-01-01T00:00:30Z'
        '403': { $ref: '#/components/responses/OriginForbidden' }
        '404': { $ref: '#/components/responses/CheckoutHidden' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    CheckoutCapability:
      type: apiKey
      in: cookie
      name: __Secure-checkout_cap
      description: |
        Capability aleatoria de al menos 256 bits emitida sólo mediante una
        cookie HttpOnly, Secure, SameSite=Strict y Path=/api/v1. El servidor
        conserva únicamente su HMAC/hash. Ausencia, expiración, alteración o
        relación ajena producen el mismo 404. No hay valores ni ejemplos.
      x-data-class: C3
      x-raw-storage: COOKIE_AND_EPHEMERAL_GUARD_MEMORY_ONLY
    ProviderWebhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: |
        Contrato abstracto P1 para una firma de evento. El nombre y algoritmo
        del adapter real permanecen bloqueados hasta SPK-02; nunca se registra
        ni se ejemplifica el valor.
      x-data-class: C3
      x-status: BLOCKED_REAL_PROVIDER
  parameters:
...
    CheckoutId:
      name: checkoutId
      in: path
      required: true
      description: Identificador opaco; la cookie relacionada sigue siendo obligatoria.
      schema: { $ref: '#/components/schemas/OpaqueId' }
      example: CHECKOUT_ID_OPAQUE
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag fuerte del checkout. Ausencia o mismatch produce ERR-09 y cero mutación.
      schema:
        type: string
        minLength: 3
        maxLength: 128
        pattern: '^"[^"]+"$'
      example: '"checkout-v7"'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Clave opaca de 16 a 128 caracteres para checkout+submit-payment.
        La clave cruda sólo existe durante la request; se persiste un HMAC.
        No se proporciona ejemplo o default.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._~-]+$'
      x-data-class: C3
      x-loggable: false
    Origin:
      name: Origin
      in: header
      required: true
      description: Origen HTTPS exacto de la SPA; se valida con allowlist y Fetch Metadata.
      schema:
        type: string
        format: uri
        maxLength: 256
    OptionalOrigin:
      name: Origin
      in: header
      required: false
      description: Si está presente debe pertenecer a la allowlist; Fetch Metadata complementa la política.
      schema:
        type: string
        format: uri
        maxLength: 256
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Hint opaco no confiable; el servidor nunca lo refleja ni registra y siempre genera un ID seguro nuevo.
      schema: { $ref: '#/components/schemas/CorrelationId' }
  headers:
    CorrelationId:
      description: Identificador opaco de la request; puede cambiar entre replays.
      schema: { $ref: '#/components/schemas/CorrelationId' }
      example: CORRELATION_ID_SAFE
    Location:
      description: URI relativa del recurso canónico creado o ya existente.
      schema:
        type: string
        format: uri-reference
        maxLength: 256
    CheckoutCapabilityCookie:
      description: |
        Cookie de capability con HttpOnly; Secure; SameSite=Strict; Path=/api/v1.
        La respuesta no se almacena en caches compartidas. El valor se omite.
      schema:
        type: string
      x-sensitive: true
      x-loggable: false
    StrongETag:
      description: ETag fuerte de la versión del checkout.
      schema:
        type: string
        pattern: '^"[^"]+"$'
      example: '"checkout-v7"'
    NoStore:
      description: Prohíbe almacenamiento de la respuesta.
      schema:
        type: string
        const: no-store
      example: no-store
...
    VaryOrigin:
      description: Evita mezclar respuestas entre orígenes.
      schema:
        type: string
        const: Origin
      example: Origin
    RateLimitLimit:
      description: Límite de la ventana aplicable.
      schema:
        type: integer
        minimum: 1
    RateLimitRemaining:
      description: Solicitudes restantes en la ventana.
      schema:
        type: integer
        minimum: 0
...
  responses:
    RequestMalformed:
      description: ERR-01 REQUEST_MALFORMED
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            malformed:
              value:
                type: https://checkout.example.invalid/problems/request-malformed
                title: Solicitud inválida
                status: 400
                detail: No pudimos leer la solicitud.
                instance: urn:problem-instance:CORRELATION_ID_SAFE
                code: REQUEST_MALFORMED
                correlationId: CORRELATION_ID_SAFE
                retryable: false
    PayloadTooLarge:
      description: ERR-01 REQUEST_MALFORMED; el body supera 16 KiB y ningún caso de uso se ejecuta.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnsupportedMediaType:
      description: ERR-01 REQUEST_MALFORMED; los comandos sólo aceptan application/json.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    OriginForbidden:
      description: ERR-02 ORIGIN_FORBIDDEN; no revela recursos.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    CheckoutHidden:
      description: ERR-03 indistinguible para inexistente, capability inválida o relación ajena.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            unavailable:
              value:
                type: https://checkout.example.invalid/problems/checkout-unavailable
                title: Checkout no disponible
                status: 404
                detail: El recurso no está disponible.
                instance: urn:problem-instance:CORRELATION_ID_SAFE
                code: CHECKOUT_NOT_FOUND_OR_FORBIDDEN
                correlationId: CORRELATION_ID_SAFE
                retryable: false
    ProductNotFound:
      description: ERR-04 PRODUCT_NOT_FOUND
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: |
        Conflicto de negocio con precondición técnicamente válida. El code
        distingue ERR-06, ERR-07, ERR-10, ERR-11 o ERR-18.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    CheckoutExpired:
      description: ERR-08 CHECKOUT_EXPIRED; sólo aplica sin pago activo.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionFailed:
      description: ERR-09 PRECONDITION_FAILED; ausencia o mismatch de If-Match.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        ETag: { $ref: '#/components/headers/StrongETag' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    FieldInvalid:
      description: ERR-05 FIELD_INVALID; no representa un conflicto concurrente.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            invalidField:
              value:
                type: https://checkout.example.invalid/problems/field-invalid
                title: Revisa los campos
                status: 422
                detail: Hay campos que requieren tu atención.
                instance: urn:problem-instance:CORRELATION_ID_SAFE
                code: FIELD_INVALID
                correlationId: CORRELATION_ID_SAFE
                retryable: false
                errors:
                  - path: installments
                    code: unsupported_value
                    reason: Elige una opción disponible.
    RateLimited:
      description: ERR-19 RATE_LIMITED; el retry sólo aplica al mismo método cuando sea seguro.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        Retry-After: { $ref: '#/components/headers/OptionalRetryAfter' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
        X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: |
        ERR-20 antes de I/O externo ambiguo. Si pudo existir envío, el recurso
        aceptado permanece consultable y se proyecta como ERR-14, no como 500.
      headers:
        Cache-Control: { $ref: '#/components/headers/NoStore' }
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    CreateCheckoutRequest:
      type: object
      additionalProperties: false
      required: [productId]
      properties:
        productId:
          $ref: '#/components/schemas/OpaqueId'
...
    CheckoutCreatedResponse:
      type: object
      additionalProperties: false
      required: [checkoutId, status, version, quote, expiresAt]
      properties:
        checkoutId:
          $ref: '#/components/schemas/OpaqueId'
        status:
          type: string
          const: DRAFT
        version:
          type: integer
          minimum: 1
        quote:
          $ref: '#/components/schemas/Quote'
        expiresAt:
          type: string
          format: date-time
    ReplaceCustomerRequest:
      type: object
      additionalProperties: false
      required: [fullName, email, phone]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          minLength: 8
          maxLength: 16
          pattern: '^\+?[0-9]{8,15}$'
      x-data-class: C2
      x-decision-status: ASSUMED_DEC-07
    CustomerResponse:
      type: object
      additionalProperties: false
      required: [customerId, checkoutId, version, fullName, email, phone]
      properties:
        customerId:
          $ref: '#/components/schemas/OpaqueId'
        checkoutId:
          $ref: '#/components/schemas/OpaqueId'
        version:
          type: integer
          minimum: 1
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          minLength: 8
          maxLength: 16
          pattern: '^\+?[0-9]{8,15}$'
      x-data-class: C2
      x-cache: no-store
...
    CheckoutResponse:
      type: object
      additionalProperties: false
      required: [checkoutId, status, version, product, quote, expiresAt]
      properties:
        checkoutId:
          $ref: '#/components/schemas/OpaqueId'
        status:
          $ref: '#/components/schemas/CheckoutStatus'
        version:
          type: integer
          minimum: 1
        product:
          $ref: '#/components/schemas/ProductPublic'
        quote:
          $ref: '#/components/schemas/Quote'
        customer:
          oneOf:
            - $ref: '#/components/schemas/CustomerResponse'
            - type: 'null'
        deliveryDetails:
          oneOf:
            - $ref: '#/components/schemas/DeliveryDetailsResponse'
            - type: 'null'
        activeTransactionId:
          oneOf:
            - $ref: '#/components/schemas/OpaqueId'
            - type: 'null'
        expiresAt:
          type: string
          format: date-time
      x-data-class: C2
      x-cache: no-store
...
    PaymentSubmissionRequest:
      type: object
      additionalProperties: false
      required: [quoteId, paymentMethodToken, installments, acceptances]
      properties:
        quoteId:
          $ref: '#/components/schemas/OpaqueId'
        paymentMethodToken:
          type: string
          minLength: 17
          maxLength: 137
          pattern: '^tok_fake_[A-Za-z0-9_-]{8,128}$'
          writeOnly: true
          description: Token C3 efímero del adapter fake de etapa 5; persistencia cero y sin ejemplo. El adapter sandbox permanece deshabilitado y requiere cambio contractual tras SPK-02.
          x-data-class: C3
          x-ephemeral: true
          x-loggable: false
          x-persisted: false
        installments:
          type: integer
          minimum: 1
          maximum: 36
        acceptances:
          $ref: '#/components/schemas/PaymentAcceptanceTokens'
      x-rejected-properties:
        - amount
        - amountInCents
        - subtotal
        - fees
        - total
        - currency
        - paymentStatus
        - providerStatus
        - providerReference
        - providerTransactionId
        - cardNumber
        - securityCode
        - expiry
...